Add controllers and update routing
