+++ controllers/error.js
@@ -1,0 +1,3 @@
+exports.get404 = (req, res) => {
+  res.render('404', { pageTitle: '404 | Not Found' });
+};
