+++ routes/shop.js
@@ -1,12 +1,9 @@
 const express = require('express');
 
-const adminData = require('./admin');
+const productsController = require('../controllers/products');
 
 const router = express.Router();
 
-router.get('/', (req, res) => {
-  const products = adminData.products;
-  res.render('shop', { prods: products, pageTitle: 'Shop', path: '/' });
-});
+router.get('/', productsController.getProducts);
 
 module.exports = router;
